Add filter_metadata config option
We had a report of an app that contains sensitive information in the
request path and the desire to filter this out. We have no system in
place to filter metadata like path and request method, as set by the
Sinatra middleware.

This change allow apps to filter out some metadata that's set by
default, like `path`, to avoid sending PII or other sensitive data,
using the `filter_metadata` config option.

Filtering is done with String based keys, like all the other `filter_*`
config options are, so the keys need to be transformed to keys
beforehand to make sure they're filtered out.

I didn't merge how we set the metadata, now it's set using
`Transaction#set_metadata` and through `sample_data` when the
Transaction is being sampled as sample data. I've left the behavior the
same as much as possible to avoid breaking things.

See also this internal discussion:
https://appsignal.slack.com/archives/CNPP953E2/p1687785270464119