Do not remove non-recursive identical values
Because the `seen` list instance was being passed along, not just
down the recursion chain, but from previous to next values, this
meant that non-recursive appearances of the same (as in same
instance) array or hash would be removed.

This commit clones the `seen` array before appending an element to
it. The clone is passed down the recursion chain, but not shared to
adjacent values.