Add behavior to not start on config file error (#991)
When the `config/appsignal.yml` file triggers an error while parsing it,
we ignore the config file source. This can cause some unexpected
behavior, like missing config and reporting it as the wrong app name.
More importantly, any filter config options in the config file are
missing and we may send PII data that's supposed to be filtered out.

We don't want to change the behavior now, as it's a breaking change. Add
a flag to opt-in to this new behavior, using the
`APPSIGNAL_INACTIVE_ON_CONFIG_FILE_ERROR` config option. Then for the
next major version we should make this the new default behavior.

If accepted, I'll make this more visible in the diagnose report.

Closes #926