Fix broken specs for Rails 7.1
Explicitly initialize the Rails error reporter for ActiveJob specs in
Sidekiq. It seems to be reporting errors that way now.

Add a helper to initialize and remove the error reporter within the same
test with a block.

By default, more tags are included, so update the tests to allow for
more than the asserted tags.