Fix deprecation for Sidekiq 8.0 error handler
Sidekiq 7.1.5 introduces a change where the Sidekiq config is not
passed in the context hash anymore, but as a separate positional
argument.

If the handler takes two arguments a deprecation warning is logged.
In 8.0 this backwards compatibility will be dropped.

Since the config is not used at all, we can simply add an optional
3rd argument that works with old and new versions.

Co-authored-by: Bart de Water <118401830+bdewater-thatch@users.noreply.github.com>