Fix error in Sidekiq probe on missing Redis info
We've seen a user report that the Redis `INFO` command on their server
doesn't report the `used_memory_rss` key. Update the probe to not error
when no value is set for that key or the key is not present. That way it
will continue to report all the other data, just not the missing data
points.

Fixes #1005