Update Makefile log file key in diagnose report
As reported in #1018 the `mkmf.log` file path needs to be changed. It
can be found in the `extensions` directory. The path can be fetched with
the `extension_dir` method on the parsed gemspec.

This `extension_dir` method will only work when installed through Ruby
gems. If installed through a local path or Git repo it will need to use
the original behavior. If no file can be found, the `path` value is not
set.

I've updated the `gem_path` method to fetch the install path from the
gemspec as well. That way we have one source for this kind of
information.

Closes #1018