Remove deprecated extension functions
I want to remove these functions from the Rust extension, but if the
Ruby gem still calls them, that would cause a new error on the next
extension update.

These functions already did nothing anymore in the Rust extension,
so not calling the extension functions loses no functionality. I did not
remove the Ruby methods, because people may still call them and we
should only remove methods in major releases.

Move the deprecation warnings to the Ruby gem and remove calls to the
extension functions.

- Part of https://github.com/appsignal/appsignal-agent/issues/425
- Deprecations logged in
  https://github.com/appsignal/appsignal-ruby/issues/667