Fix Ruby and JRuby versions in install report
The Ruby version for JRuby was reported as `-p` because it didn't have
the `RUBY_PROGRAM_VERSION` and `PATCHLEVEL` values from RbConfig.
Instead use the `ruby_version` value that seems available on both
implementations.

For JRuby, also track the implementation version so we can see which
JRuby is being used. This will require a server update to show this
field.