Test extension during JRuby install (#1031)
When installing the extension for JRuby we didn't actually test if the
extension was working or not. The `install.report` would report the
extension was installed successfully if the artifact archive was
downloaded and extracted. We didn't test if the extension functions
could be "attached" using FFI.

Load the JRuby extension during installation to test if the functions
could be attached without issues.

When it fails, raise the error rather than logging it, so that the
installer can detect and rescue the error to store the installation
failure in the `install.report` file.