Rename the internal logger method
To avoid confusion and potential misuse of the AppSignal internal logger
rename `Appsignal.logger` to `Appsignal.internal_logger`.

We are worried some people may be using `Appsignal.logger` to log
messages to our logging feature, which are two different things.
Hopefully the name "internal" in the method call should make it clear
it's not meant to be called directly.

[skip changeset] This is the change of a private API. People should
already not be calling it.

Closes #1033