Update supported gems in environment metadata
Update which gems we support in the environment metadata system, since
this was last updated. This way we have a better idea of what versions
to support of those gems.

I've also added more Rails gems that could be used outside of Rails.
That way we don't miss those gems when they're not part of a Rails app.

[skip changeset] as it's not a public facing change.