Monitor serving of Rack response bodies
Some work might be getting done within a Rack response body. For example, when
ActionController::Streaming is used, or when a Rack app elects to stream a response.

The Rack SPEC doc actually defines the behavior in sufficient detail to wrap this
into the same Appsignal transaction.

The Rails instrumentation may also inherit from the more generic one as most
methods and constants are shared then