Complete txn without body too
If there is an exception within `call` of a Rack app, the body won't be returned. And that will be the case 99% of the time. If that happens we need to complete the transaction early to ensure data is sent across, as there is no BodyWrapper to complete it for us.