Allow gem users to customize behavior on errors in ActiveJob
Considering the following setup using shoryuken, appsignal and our own custom queue adapter (raven):

```
Shoryuken.configure_server do |server|
  server.server_middleware do |chain|
    chain.clear
    chain.add Appsignal::Hooks::ShoryukenMiddleware
    chain.add Raven::ShoryukenMiddleware
  end
end
```

The `Raven::ShoryukenMiddleware` will perform the job and if it errors more than X times it will exhaust the job and report
the error to appsignal. We 'swallow' the exception so the Appsignal::Hooks::ShoryukenMiddleware will not report is as well.
This doesn't work though since the `::Appsignal::Hooks::ActiveJobHook::ActiveJobClassInstrumentation` already reported
the error during the performing of the job since it overrides active_job's execute method.
To allow more fine-grained programmatic control the actual setting of the error is moved to a method which
gem users can override and customize. The method name is prefixed with `appsignal` to minimize the change
for unwanted naming collisions.