Clear in memory log on logger start
The `@in_memory_log` would continue to keep the messages in memory that
were logged during the gem start, before a logger was initialized. This
keeps memory allocated while apps are running.

This issue was reported in #645, where it also kept logging those in
memory logged messages whenever a process was forked and AppSignal
started.

Clear the in memory log so it uses less memory and doesn't result in
stale logs being logged on process forks.

Fix issues with the test suite where state on the `Appsignal` module
persisted between tests by removing the `@in_memory_log` variable and
clearing the `@internal_logger` variable.

Closes #645