Start probes registered after thread is started
When the AppSignal gem starts, it is no longer possible to register
probes. If AppSignal is started by one of our framework integration, it
may be difficult to register a probe before then.

It was required that people manually called `Appsignal::Probes.start`
again.

This change auto initializes newly registered probes if the minutely
probes thread has already started. It will then be included in the next
iteration of the thread's loop.

Closes #812