Report errors for Active Job < 7.1 with discard
Report all errors for Active Job < 7.1 when `activejob_report_errors` is
set to `discard`. Previously, it would ignore errors reported when using
this incompatible option.

Use the config's override functionality for this to override the
`discard` value to `all` when Active Job < 7.1 is used.

I wanted to do this in the integration itself, but I cannot for the life
of me test Active Job in the way needed with the inline or test
adapters. Up side, we can now see these kinds of misconfigurations in
the diagnose report.