Track the number of executions of Active Job jobs
When a job is retried, the execution count goes up every time. To give
more insights into how many times a job has been retried, track the
executions as a tag.

Executions are zero when the job is run the first time for some reason,
so +1 the value so it makes more sense to a human being reading it.
Calling `to_i` on the job metadata just to make sure we don't error on
any weird value type.