Use the Ruby debug gem instead of pry
Switch to using debug instead of pry for debugging.
I've found myself use debug all the, but it not being present in the
development tooling of the gem is annoying.

Remove the default require for pry from the spec helper now that it's
removed. I don't need the debug gem to be loaded all the time either as
my snippet for it includes the require statement.

[skip changeset]