Do not use `trace` on internal logger
The internal logger does not support `trace`. Use `debug` instead.

Since the heartbeat implementation intentionally swallows all errors
and logs them, issues such as this would go unnoticed by the test
suite. Add assertions on the messages that are expected to be
logged to ensure that the implementation is actually working as
expected.