Report request response status for Rails apps
We can now report the response status as a `response_status` tag and
metric. Previously, due to the position of the Rails middleware, this
was unreliable and wouldn't always report the real status code if other
middleware in the stack that are run after ours returned another status.

This is part of issue #183, but only implements it now for Rails apps
and apps using the `Appsignal::Rack::EventHandler` manually. We'll need
to update other instrumentations for Rack, Sinatra, Padrino, Grape, etc.
to use this new EventHandler, so they can also benefit from this new
response status tag and metric.