Ignore requests without a Rails controller
Don't set the action from the AppSignal EventHandler. This
`request.path` is not normalized and we would report too many unique
actions. This means the EventHandler can't be used on its own without
calling `Appsignal.set_action`.

We do this so that for Rails (the only instrumentation that uses the
EventHandler right now), will not report any samples for requests for
assets. Those asset requests will not have a controller, the
RailsInstrumentation middleware will not report an action name, and the
extension will drop the transaction as no action name is set.