Fix compatibility with Rack 3.1.0
Rack 3.1.0 broke our assertion on the default sample data. Previously a
`rack.input` object was always set by the env returned by
`Rack::MockRequest`. Since this commit this is no longer required:
https://github.com/rack/rack/commit/54368a0981c46c2c62137e47795b38de50dabefb

I could add it back so there's an input by default again, but if we were
to implement this now from scratch we wouldn't add this assertion.
Instead, I'm leaving it out.

This test doesn't assert that the CONTENT_LENGTH environment is required
to be set. It just asserts a couple environment values are present.

[skip changeset]