Fix logger error call with exception object
It was reported that calling `Rails.logger.error` with an exception
object raises an ArgumentError. This comes from the extension, which
expects the log message to be a String, not an Exception object.

```ruby
Rails.logger.error StandardError.new("StandardError log message")
```

The log message for the exception will only be the exception class and
the message. The backtrace is omitted because it would just squash the
multiple lines from the backtrace on one line, making it difficult to
read.

Fixes #1095