Report Sidekiq errors when the job is discarded
Add the `sidekiq_report_errors` config option. If set to `discard`, it
will only report the error if the job is discarded, or "dead" in Sidekiq
terminology.

This follows the same behavior as the `activejob_report_errors` config
option.

To make this behavior work, it uses a Sidekiq death handler to report
the error when the job is discarded. This death handler is run before
the error handler, so the error handler can stay responsible for
completing the transaction.

Death handlers were introduced in Sidekiq 5.1. Older versions will
report all errors for jobs.

Part of #1091