Fix Sinatra instrumentation causing boot loop
I noticed in our test setup that when I added a Sinatra app to the Rails
app, and loaded the Sinatra instrumentation as specified in our docs it
would get stuck in a boot loop. This was caused by the two different
configs overwriting each other every time.

Prevent this issue by skipping loading the AppSignal config if it's
already active.