Use AbstractMiddleware for Rails middleware
Use our new AbstractMiddleware as a basis for the Rails instrumentation
middleware. This moves all responsibility for handling nested
instrumentation middleware to the AbstractMiddleware.

The Rails middleware can now focus on only the part it needs to do: the
action name and some additional metadata.

The RailsInstrumentation middleware previously didn't create an event,
but the AbstractMiddleware does do that now. It will report a new
`middleware.rails` event now.

I didn't add a changeset for this change as it's mostly an internal
refactor.