Add Transaction set_params(_if_nil) methods
Add methods to set the parameters on the Transaction. This replaces and
deprecates the Transaction `params=` writer method that was used before
this.

We need the `_if_nil` variant for the AbstractMiddleware to only write
the parameters if they're not already set by the app it's instrumenting.
Currently it tries to set the parameters before the app handles the
request, but this is unreliable and doesn't always report the parameters
of the request.

The `_if_nil` variant isn't perfect, because it doesn't consider
whatever parameters could be fetched from the request object, because it
now skips this. I can improve this later if necessary.

Part of #1108