Fix parameter reporting for Rack and Sinatra apps
The AbstractMiddleware tried to set the request parameters before the
app handled the request. This could cause the request object to end up
in a weird state where, if some other middleware (like a JSON payload
parsing middleware) hadn't parsed the request payload yet, the
parameters were reported as empty.

Move the parameter setting to the "after" hook so that it doesn't try to
read the parameters before they're parsed. Use the new
`Transaction#set_params_if_nil` so that it doesn't overwrite any custom
parameters set by the app.

Fixes #1108