Fix error in EventHandler on request with error
When a request has an error, it would give the `on_finish` callback a
response object that is `nil`.

This would break when we called `response.status`, and it would fail to
complete the transaction.

Make sure the transaction is always completed when it's handling a
transaction, by putting the outside the safety block which fetches all
the request and response metadata.

Check if the `response` object is `nil` before calling other
methods on it to the response status tag and metric. This way we don't
raise an error.

Luckily no integrations of ours that have been published yet would run
into this behavior. The Rails integrations uses the EventHandler, but
has its own error reporting.

This would only be a problem for people who use the EventHandler
directly, or when we move more integrations over to use
AbstractMiddleware subclasses.