[WIP] Rename heartbeats to cron check-ins
This is pending confirmation of the specific terms of the rename,
as well as the shape of the new endpoints and so on. The specific
names used here are placeholders.

Invoking the previous method and module names raises a deprecation
warning the first time it takes place, and forwards the call to the
new method and module names.