Fix error reporting for Rails instrumentation
Updating the RailsInstrumentation middleware to inherit from the
AbstractMiddleware broke the error reporting for the middleware.
It relied on the EventHandler to report errors, but before it gets
there, Rails's exception handling middleware rescues the error and shows
an error page instead.

Add a config option to the AbstractMiddleware to configure if it should
report errors or not. Enable it for the RailsInstrumentation to restore
the error reporting functionality.