Support nested Grape apps
Refactor the Grape middleware to inherit from the AbstractMiddleware
class. I've removed the inheritance on the Grape::Middleware::Base. We
don't rely on the base Grape middleware.

By using an AbstractMiddleware subclass for Grape we support Grape apps
nested in Sinatra, Rails and other apps already using AppSignal
instrumentation middleware.

To keep the `grape.skip_appsignal_error` request env logic to work, I've
made the `report_errors` option accept a proc and call it if it's
callable to determine per request if the error should be reported.