Deprecate Appsignal.start_logger
We always tell people to call both `Appsignal.start` and
`Appsignal.start_logger`. Let's make it easier for everyone and start
the logger when AppSignal is started?

For the one scenario where the logger needs to be called separately in
our gem, we can call the private `_start_logger` method.