Fix in memory logger not logging all messages (#1121)
When the in-memory logger is active (before AppSignal is started), it
ignores the log level and only logs messages of level info or higher. On
AppSignal start, all debug messages were lost, and we log a couple of
debug messages during boot.

I first thought of setting the in-memory logger level to debug, but this
results in the log file containing debug messages, even if the logger
eventually wasn't set to the debug log level.

I added a new class called IntegrationMemoryLogger that keeps the log
messages in a Hash so that it can figure out later what log messages it
should copy to the initialized logger. Now, the debug messages are
logged when the 'real' logger is initialized with the debug log level.

[skip changeset]