Add process name and ID tags to GVL metrics
When reporting the GVL metrics, which measure the time spent by the
Ruby VM in its global VM lock, and as such are per-process metrics,
add the process name and process ID as tags, alongside the existing
hostname tag.

As the process name can change during the lifetime of the process,
store it when initialising the GVL probe, to prevent potentially
reporting the same metric with different tag combinations.

To attempt to account for custom process names, keep only the first
word of the process name, after removing path prefixes from it.

Fixes #1106.