Report response status when there is no response
When there is no response object given to the EventHandler's `on_finish`
callback, we can expect an unhandled error has occurred.

When have also received an exception in the `on_error` callback, report
the response status as 500. This solution is not perfect. It may
actually be another response status if something else above the
EventHandler handles the exception. A setup we don't recommend.