Fix support for Hanami 2.1
In Hanami 2.1 the `sessions_enabled` keyword argument was renamed to
`session_enabled` for the `Hanami::Action::Request` class. Upon closer
inspection, the `Hanami::Action::Request` class does nothing we need, so
let's remove our dependency on it.

To fetch the request parameters, directly ask the Hanami Action request
parameters class. This is (also) a private API, but limits the surface
area in which things can break in the next release. I'll continue
looking at a better to fix this.

Fix specs for the Hanami middleware options. They added another item to
the registered middleware list.