Improve Hanami integration testing
Improve testing for the Hanami integration by making sure the it
prepends the module in every scenario. That means we're not really
prepending the module to Hanami::Action in the tests, because that would
make testing it multiple times impossible. It's not possible to
"unprepend" a module.

I've updated the prepended module test to prepend the module to an
Action class. First it subclasses the Hanami Action fixture using an
anonymous class (using `Class.new`). Then it prepends the module to that
anonymous class to see if the behavior works.