Don't start again in Hanami integration
When AppSignal is already active, do not start AppSignal again.
This is a good precaution as it prevents boot loops, when AppSignal is
started twice with different configurations.

This will improve support for nested Hanami applications, when they're
mounted in another frameworks we support, like Rails or Sinatra.

This is similar to PR #1105

Part of #329