Remove default abstract instrumentation event
In PR #1124, I added functionality to not report an instrumentation
event for the Rails instrumentation middleware.

The AbstractMiddleware shouldn't be used on its own, but it did have a
default for reporting instrumentation events. Remove that default so it
doesn't report an event by default.

[skip changeset] It's an internal change to the AbstractMiddleware and
no one should be using that directly.