Deprecate our GenericInstrumentation middleware
Deprecate the Appsignal::Rack::GenericInstrumentation in favor of the
Appsignal::Rack::InstrumentationMiddleware. This other middleware
doesn't set the action name to "unknown" by default.

The GenericInstrumentation is no longer loaded automatically. This helps
with printing a deprecation warning when it is used.

Update the Grape specs which I copied this from to not mention the
Minutely constant, which I copied it from there.