Support nested webmachine apps
Apply the same logic as we have in the AbstractMiddleware for webmachine
apps. When a parent transaction is active, use that and don't create a
new one. More importantly, don't close the active transaction.

The webmachine instrumentation can't use the AbstractMiddleware, or any
middleware, as this discouraged by the webmachine project. In practice I
don't see this happening, but if someone happens to add an AppSignal
EventHandler to the middleware stack in `config.ru`, we now support it.

Part of #329