Use Rack middleware for Padrino instrumentation
Refactor the Padrino instrumentation to use as much existing Rack
middleware for instrumentation as possible.

Padrino is a framework built on Sinatra. We can reuse the Sinatra
instrumentation middleware to instrument Padrino apps. In combination
with the EventHandler we have 99% of the instrumentation done.
To differentiate between Padrino and Sinatra events I've configured the
Sinatra middleware to report the instrumentation event as a Padrino
event.

The only thing we do need a monkeypatch for is the action name for
Padrino. Everything else can be removed from the monkeypatch. The
request metadata and exceptions are reported by the Sinatra middleware.

I've moved all the plugin initialization to the `Padrino.before_load`
callback. It seems to initialize all the things we need before this
time: app path and environment.