Fix Padrino instrumentation causing boot loop
Apply the same boot loop fix as for Sinatra for Padrino from PR #1105.

For Sinatra apps noticed in our test setup that when I added a Sinatra
app to the Rails app, and loaded the Sinatra instrumentation as
specified in our docs it would get stuck in a boot loop. This was caused
by the two different configs overwriting each other every time.

Apply the same fix for Padrino by checking first if AppSignal has
already started and skipping the auto start.