Add Appsignal.report_error helper
To help alleviate issues with confusing around when to use
`Appsignal.set_error` and `Appsignal.send_error`, introduce the
`Appsignal.report_error` helper.

This helper will always report the error. If a transaction is active,
it reports the error on the current transaction. If no transaction is
active, it will create a new transaction and report the error this way.

It should allow apps to call this helper method in all contexts: with
and without an active transaction.

Closes #176
Related to https://github.com/appsignal/integration-guide/issues/167