Add monitor instrumentation helpers
Add new "create a transaction" helpers to replace the
`monitor_transaction` and `monitor_single_transaction` helpers.
These helpers do too much and are confusing to use. We didn't end up
using them ourselves anymore in our integrations.

These new helpers are based on what we most commonly do in the Ruby gem
to instrument blocks of code.

These helpers don't accept adding metadata to the transaction using an
`env` argument. This can now be set using the instrumentation helpers
like `set_action`, `set_params`, `set_custom_data`, etc.

If an app wants to track blocks of code with an instrumentation event,
`Appsignal.instrument` needs to be called in the `Appsignal.monitor`
block.

If this helper were to both create a transaction and an instrumentation
event the arguments would get complicated and it becomes unclear what
it's meant to be used for.

These helpers support nested transaction in such a way that it won't
break anything, but the namespace and action arguments are ignored when
a parent transaction is active.